Refactor: Break down MajorDetails.tsx
Refactor MajorDetails.tsx into smaller, more manageable components to improve code organization and maintainability.
